Extract broker UDP reactor module

Copilot-Session: b5a1a347-37a8-4246-8bbc-306590921475
